Add type information to the grammar.
The required syntax now looks something like this:

    (ident type[:label])

where "ident" is a string identifier, and "type" is a primitive
or another identifier, and "label" is an optional label (preceded
by a colon).

For example:

    (small_int 97 uint8)
    (named_int 98 int:foo)

For integer literals, the proper type is automatically inferred
unless explicitly specified. The currently accepted primitive
types are uint8, uint6 and int.

Note that the "int" will be truncated internally to 31-bits
(although serialization will be done in 32-bits); actual 32-bit
support will come soon.