Join common cycles together as one cycle.
Cycle dependencies are transitive, so cycles of A B C A and
D E C D share the same dependencies, since 'C' is a common element.

The basic algorithm compares two sets (of strings), unifies them
if the intersection is nonempty, and adds into a set-of-sets.

If the new set-of-sets is equal to the input set-of-sets, then
we've converged. Otherwise, recurse with the new set-of-sets.